# doc/openapi.yaml
swagger: '2.0'
info:
  contact: {}
  title: Breezy API
  description: API agrégée de la plateforme Breezy (réseau social microservices).
  version: '1.0'
paths:
...
  /users/follow-requests/{followerId}/accept:
    post:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Accepter une demande d'abonnement
      parameters:
      - type: string
        description: ID du demandeur
        name: followerId
        in: path
        required: true
      responses:
        '200':
          description: 'status: accepted'
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /users/follow-requests/{followerId}/reject:
    post:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Rejeter une demande d'abonnement
      parameters:
      - type: string
        description: ID du demandeur
        name: followerId
        in: path
        required: true
      responses:
        '200':
          description: 'status: rejected'
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
  /users/me/follow-requests/outgoing:
    get:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Demandes d'abonnement sortantes (envoyées par moi)
      responses:
        '200':
          description: Liste d'IDs
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
  /users/me/followers/{id}:
    delete:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Retirer un abonné de sa liste
      parameters:
      - type: string
        description: ID de l'abonné à retirer
        name: id
        in: path
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
...
  /users/{id}/follow:
    post:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Suivre un utilisateur
      parameters:
      - type: string
        description: ID de l'utilisateur à suivre
        name: id
        in: path
        required: true
      responses:
        '200':
          description: 'status: following|requested'
          schema:
            type: object
            additionalProperties:
              type: string
        '400':
          description: Bad Request
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
    delete:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Ne plus suivre un utilisateur
      parameters:
      - type: string
        description: ID de l'utilisateur
        name: id
        in: path
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
  /users/{id}/followers:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Lister les abonnés d'un utilisateur
      parameters:
      - type: string
        description: User ID
        name: id
        in: path
        required: true
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.User'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /users/{id}/following:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Lister les abonnements d'un utilisateur
      parameters:
      - type: string
        description: User ID
        name: id
        in: path
        required: true
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.User'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
definitions:
...
  models.User:
    type: object
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
      username_changed_at:
        description: 'UsernameChangedAt : date du dernier changement effectif du handle

          (nil = jamais changé). Sert de base au cooldown de changement de

          username (cf. service). Symétrique de display_name_changed_at (profil).'
        type: string
...
host: localhost:8080
basePath: /